fix narrowing conversion

+++ src/Core/GuiControlDropDownBox.cpp
@@ -90,7 +90,7 @@ GuiControlButton* GuiControlDropDownBox::AddOption(const std::string& optionText
 	int id = app->guiManager->GetGuiLastId() + 1;
 
 	// Create a new GuiControlButton with the given text
-	GuiControlButton* newOption = (GuiControlButton*) app->guiManager->CreateGuiControl(GuiControlType::BUTTON, id, optionText.c_str(), {bounds.x, bounds.y + (options.size() + 1) * 30, bounds.w, 30}, observer);
+	GuiControlButton* newOption = (GuiControlButton*) app->guiManager->CreateGuiControl(GuiControlType::BUTTON, id, optionText.c_str(), {bounds.x, static_cast<int>(bounds.y + (options.size() + 1) * 30), bounds.w, 30}, observer);
 
 	newOption->state = GuiControlState::DISABLED;
